+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 Changed
 - Reworked `Instruction` object
 - Reworked `Run.instructions` to return a Dataframe of `Instruction` objects
+- `Aliquot` object has been reworked into Container object as an `aliquots` property
+
+Removed
+- `Resource` object has been removed from the library as its currently unused
 
 Fixed
 - Change check for ImagePlate to be more generic